Crate tc3xx_service

Startup code and minimal runtime for TriCore(TM) TC162 microcontrollers

This crate contains all the required parts to build a no_std application (binary crate) that targets a TC3xx AURIX(TM) microcontroller family.

ARCHITECTURE: TC162

UC FAMILY: TC3xx

Features

This crate takes care of:

C/C++/RUST runtime environment startup file.

To initialize ABI registers

To initialize CSA context area

To initialize Stack area needed for ‘C’ code

To initialize ECC in RAM memories (not done for AURIX)

To clear BSS (not-initialized) data

To initialize DATA (initialized) data

To initialize C++ global object instances

To provide startup hooks for user defined code

Multicore Startup

Crt0 startup on multicore systems is driven by a configuration provided by a corresponding linker file. Configuration table contains one entry for each Core present in uC

Startup Hooks

Crt0 provides two startup hooks the user can use to implement system dependent early HW initialization.

‘Crt0PreInit()’ hook called

after STACK area and CSA list on given core is initialized

before Clear and Copy table operation is executed ! HW initialization code cannot rely on global data (BSS or DATA)

‘Crt0PostInit()’ hook called

when all startup initialization is finished (data can be used)

before calling shared_main() application entry point

C++ support

The crt0 startup code provides a default global object initialization based on ctor table registered per core. The format of ‘ctor’ table expected: ctor_setion { CTOR_LIST = .; LONG((CTOR_END - CTOR_LIST) / 4 - 2) KEEP (*(.ctors)); LONG(0); CTOR_END = .; } By default, the linker gathers all C++ module ctor entries into one common ctor table. In case of multicore object initialization, the user must provide their system of CTOR initialization.

Application specific crt0 Routines Implementation]

Startup routines are implemented as WEAK functions to allow application-specific replacement in case of need.

Crt0BssInit

Crt0DataInit

Crt0CtorInit

Crt0PreInit

Crt0PostInit

FILE HISTORY:

V1 01.2023 MF

Base reference example

Modules

asm_calls

Module has some functions which provide some low level functions for Aurix microcontrollers. The functions use inline assembly.

interrupt

The module contains the interrupt enumeration and the handler assignment macro.

Macros

entry

Macro defines a main entry point for embedded project.

interrupt

Assigns a handler to an interrupt

post_init

Macro defines a post_init entry point for embedded project.

pre_init

Macro defines a pre_init entry point for embedded project.

Functions

_exit⚠

Declaration of _exit function, which is implemented in inline assembly.

abort⚠

This function should be called on error.